+++ src/Console/PrintSchemaCommand.php
@@ -39,16 +39,15 @@ SIGNATURE;
 
                 return;
             }
+
             $filename = self::GRAPHQL_FEDERATION_FILENAME;
             $schemaString = FederationPrinter::print($schema);
+        } elseif ($this->option('json')) {
+            $filename = self::JSON_FILENAME;
+            $schemaString = $this->toJson($schema);
         } else {
-            if ($this->option('json')) {
-                $filename = self::JSON_FILENAME;
-                $schemaString = $this->toJson($schema);
-            } else {
-                $filename = self::GRAPHQL_FILENAME;
-                $schemaString = SchemaPrinter::doPrint($schema);
-            }
+            $filename = self::GRAPHQL_FILENAME;
+            $schemaString = SchemaPrinter::doPrint($schema);
         }
 
         if ($this->option('write')) {