+++ tests/database/migrations/2018_02_28_000002_create_testbench_users_table.php
@@ -9,19 +9,21 @@ class CreateTestbenchUsersTable extends Migration
     public function up(): void
     {
         Schema::create('users', function (Blueprint $table): void {
-            $table->increments('id');
+            // Mostly following https://github.com/laravel/laravel/blob/master/database/migrations/2014_10_12_000000_create_users_table.php
+            // but marking some fields nullable to simplify tests
+            $table->increments('id'); // TODO use id() in later Laravel
             $table->string('name')->nullable();
-            $table->string('email')->nullable();
+            $table->string('email')->unique()->nullable();
+            $table->timestamp('email_verified_at')->nullable();
             $table->string('password')->nullable();
-            $table->string('remember_token')->nullable();
+            $table->rememberToken();
+            $table->timestamps();
 
             $table->unsignedInteger('company_id')->nullable();
             $table->unsignedInteger('team_id')->nullable();
 
             $table->unsignedInteger('person_id')->nullable();
             $table->string('person_type')->nullable();
-
-            $table->timestamps();
         });
     }